[Parkour] Improve ledge grabbing
Grabbing now only occurs when there's empty space a little above where you're trying to grab, rather than if the tile above the one you're trying to grab is non-solid. This means you can no longer grab super low on a ledge and you can grab walls that have a slope above them. Also adjusted the rect we check for the initial wall collision with to avoid being able to grab while standing on a slope.